swapped over to using longs for media amounts everywhere. Added CircleMishapEnv. Setup CircleCastEnv more.

+++ Common/src/main/java/at/petrak/hexcasting/api/addldata/ADHexHolder.java
@@ -15,7 +15,7 @@ public interface ADHexHolder {
     @Nullable
     List<Iota> getHex(ServerLevel level);
 
-    void writeHex(List<Iota> patterns, int media);
+    void writeHex(List<Iota> patterns, long media);
 
     void clearHex();
 }
